# docs/svc_controle_coherence/static/openapi.yaml
openapi: 3.0.0
info:
  title: API Contrôle de Cohérence DPE / Audit
  description: >
    API REST permettant la validation, la traduction et le contrôle de cohérence
    des fichiers XML relatifs aux Diagnostics de Performance Énergétique (DPE)
    et aux Audits énergétiques.
  version: "1.24.0"
paths:
...
  /controle_coherence_test_1_janvier_2026:
    post:
      summary: Contrôle de cohérence DPE avec PEF électricité 1.9
      description: Permet de tester le moteur de cohérence en appliquant le nouveau facteur de conversion (PEF = 1.9).
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
      responses:
        '200':
          description: Résultat du contrôle simulé (PEF = 1.9)
  /controle_coherence_audit:
    post:
      summary: Contrôle de cohérence d’un Audit énergétique
      description: Lance la procédure de validation et de cohérence sur un fichier XML Audit.
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
              example: "<audit>...</audit>"
      responses:
        '200':
          description: Résultat du contrôle de cohérence Audit
          content:
            application/json: {}
        '400':
          description: XML invalide ou manquant
        '500':
          description: Erreur interne du moteur
  /controle_coherence_audit_test_1_janvier_2026:
    post:
      summary: Contrôle de cohérence Audit avec PEF électricité 1.9
      description: Permet de tester le moteur de cohérence pour les audits avec le nouveau facteur de conversion.
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
      responses:
        '200':
          description: Résultat du contrôle simulé (PEF = 1.9)
...